+++ CHANGELOG.md
@@ -7,6 +7,10 @@ All notable changes to this project will be documented in this file.
 ### Fixed
 - Fixed panic in openutils.ObjectLen.
 
+### Added
+- Added ConsoleLogger shared openutil. ConsoleLogger implements simple logger to `server-console.txt` file.
+- Added `openutils.ConvertTableToJson` and `openutils.ConvertJsonToTable` functions to `openutils.lua`.
+
 ## [v0.5.1] - 2024-06-16
 ### Fixed
 - Fixed the ability for the administrator to interact with a vehicle that is in another player’s safehouse. Switch Vehicles Mechanics cheat to turn it on or off.